Use cached mamehist when server is over quota.
Updates #120

+++ ds/mame.go
@@ -106,6 +106,10 @@ func updateMAMEDB(version, p string) error {
 	if resp.StatusCode != http.StatusOK {
 		return fmt.Errorf("got %v response", resp.Status)
 	}
+	if version != "" && resp.StatusCode == http.StatusTooManyRequests {
+		log.Printf("WARN: Using cached MAME History. Server over quota.")
+		return nil
+	}
 	dbp := filepath.Join(p, mameDBName)
 	err = os.RemoveAll(dbp)
 	if err != nil {
